create payment methods items
