+++ lib/features/check_out/presentation/widgets/checkout_details.dart
@@ -63,48 +63,3 @@ class CheckoutDetails extends StatelessWidget {
 
 
 
-class PaymentMethodItem extends StatelessWidget {
-  const PaymentMethodItem({
-    super.key,
-    required this.isActive,
-    required this.image,
-  });
-
-  final bool isActive;
-  final String image;
-  @override
-  Widget build(BuildContext context) {
-    return AnimatedContainer(
-      duration: const Duration(milliseconds: 300),
-      width: 103.w,
-      height: 62.h,
-      decoration: ShapeDecoration(
-        shape: RoundedRectangleBorder(
-          side: BorderSide(
-            width: 3.w,
-            color: isActive ? ColorManager.kPrimaryColor : Colors.grey,
-          ),
-          borderRadius: BorderRadius.circular(15),
-        ),
-        shadows: [
-          BoxShadow(
-            color: isActive ? const Color(0xFF34A853) : Colors.white,
-            blurRadius: 4,
-            offset: const Offset(0, 0),
-            spreadRadius: 0,
-          )
-        ],
-      ),
-      child: Container(
-        decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(15), color: Colors.white),
-        child: Center(
-          child: SvgPicture.asset(
-            image,
-            width: 60.w,
-          ),
-        ),
-      ),
-    );
-  }
-}
